ml_worker/requirements.txt
pika
transformers
logging
requests
python-jose